+++ src/propagate/propagate.jl
@@ -1,133 +1,120 @@
-"Object to keep the state of the propagation process"
-mutable struct Propagation{T}
-  pending::Set{SrcValue}
-  value_content::Dict{SrcValue, T}
-  touched_arrows::Set
-  sprtvals::Dict{SubPort, T}
-  carr::CompArrow
-  propagator::Function
-  function Propagation{T}(carr::CompArrow,
-                          seed::Dict{SubPort, T},
-                          propagator::Function) where T
-    p = new{T}()
-    p.carr = carr
-    p.sprtvals = seed
-    p.pending = Set(SrcValue(sport) for sport in keys(seed))
-    p.touched_arrows = Set()
-    p.value_content = Dict(SrcValue(sport) => content
-                          for (sport, content) in seed)
-    prologue, p.propagator = propagator()
-    prologue(p)
-    p
-  end
-end
-"""
-Propagate values around a composite arrow.
-# Arguments:
-  `carr`: Composite Arrow to propagate through`
-  `sprtvals`: Mapping from `SubPort` to some value of type `T`
-  `propagators`: functions
+"Name (e.g. :value, :size, ...) to abstract value (ConcreteValue, Size, ...)"
+AbValues = Dict{Symbol, Any}
 
-- Principles
--- Every arrow can propage one or more than one value
--- It can propagate that information to any one of its ports
--- If there is already values at those ports then some kind of conflict
-  resolution is necessary
--- The propagation data can depend on different types of values at differnet
-   ports
-# Returns:
-  port->value map for all ports in composite arrow
-"""
-function propagate!{T}(carr:: CompArrow,
-       sprtvals::Dict{SubPort, T},
-       propagator::Function)::Dict{SubPort, T}
-  propagation = Propagation{T}(carr, sprtvals, propagator)
-  while !(isempty(propagation.pending) && isempty(propagation.touched_arrows))
-    while !isempty(propagation.pending)
-      value = pop!(propagation.pending)
-      propagate!(value, propagation)
-    end
-    propagate_through!(propagation)
-  end
-  propagation.sprtvals
-end
+"Mapping from `port_id` to abstract values"
+IdAbValues = Dict{Int, AbValues}
 
+"Abstract TraceValues assigns abtract values to TraceValues"
+AbTraceValues = Dict{TraceValue, AbValues}
 
-"filter already processed values"
-function propagated_values{T}(sarr::SubArrow, prop::Propagation{T})
-  values = Set(all_values(sarr))
-  values ∩ keys(prop.value_content)
-end
+# FIXME: This is quite a few layers of misdirection
+"Get `sprt` in `val_abval` assuming `sprt` is on root"
+Base.get(val_abval::Dict{TraceValue, AbValues}, sprt::SubPort) =
+  val_abval[trace_value(sprt)]
 
+has(sm::Symbol) = prop -> haskey(prop, sm)
 
-"filter not yet processed values"
-function unpropagated_values{T}(sarr::SubArrow, prop::Propagation{T})
-  values = Set(all_values(sarr))
-  setdiff(values, keys(prop.value_content))
-end
+"All abstract evaluators of `arr`"
+all_abinterprets(arr::Arrow)::Set{Function} = Set(vcat(accumapply(abinterprets, arr)...))
 
-"add values to the propagation"
-function add_content!{T}(prop::Propagation{T}, sport::SubPort, content::T)
-  prop.sprtvals[sport] = content
-end
-
-"add values to the propagation while including `sub_arrow`"
-function add_content_arrow!{T}(prop::Propagation{T}, sport::SubPort, content::T)
-  add_content!(prop, sport, content)
-  push!(prop.touched_arrows, sub_arrow(sport))
-end
-
-"state that a value needs to propagate some content"
-function add_pending!{T}(prop::Propagation{T}, value::SrcValue, content::T)
-  prop.value_content[value] = content
-  push!(prop.pending, value)
-end
-
-"""This function is the basic way in which content is propagated thrhough an
-  arrow: all Values connected to the arrow will have the same content"""
-function same_content_propagator()
-  function f(sarrow::SubArrow, prop::Propagation)
-    selected_value = first(propagated_values(sarrow, prop))
-    content = prop.value_content[selected_value]
-    unpropagated = unpropagated_values(sarrow, prop)
-    foreach(value-> add_pending!(prop, value, content), unpropagated)
-  end
-  x->nothing, f
-end
-
-"""This function allows the information to jump over the arrows. The idea is
-  that we may provide specialised functions for the different kind of
-  arrows/values. For instance, if we are propagating a shape, the behavior of a
-  `MatrixMultArrow` is very different than `AddArrow` """
-function propagate_through!(prop::Propagation)
-  f = sarrow->prop.propagator(sarrow, prop)
-  foreach(f, prop.touched_arrows)
-  prop.touched_arrows = Set()
-end
-
-"check conflict during the propagation"
-function check_conflict{T}(prop::Propagation{T}, sport::SubPort, content::T)
-  if prop.sprtvals[sport] != content
-    throw(DomainError(msg))
-  end
-end
-
-"propagate the content of a value to all its subports"
-function propagate!{T}(value::SrcValue, prop::Propagation{T})
-  content = prop.value_content[value]
-  for sport in sub_ports(value)
-    if haskey(prop.sprtvals, sport)
-      check_conflict(prop, sport, content)
-    else
-      add_content_arrow!(prop, sport, content)
+"Cycle abstract evaluator of `arr` until a fixed point is reached"
+function cycle_abinterprets(arr::PrimArrow, abvals::IdAbValues)
+  abinterprets = all_abinterprets(arr)
+  atfixedpoint = false
+  while !atfixedpoint
+    atfixedpoint = true
+    for abinterpret in abinterprets
+      subabvals::IdAbValues = abinterpret(arr, abvals)
+      # Do resolution on each Value
+      for i in keys(subabvals)
+        if i in keys(abvals)
+          resabval = meetall(abvals[i], subabvals[i])
+          if resabval != abvals[i]
+            abvals[i] = resabval
+            atfixedpoint = false
+          end
+        else
+          atfixedpoint = false
+          abvals[i] = subabvals[i]
+        end
+      end
     end
   end
+  abvals
 end
 
-"""propagate the content of a value to all its subports,
-  using the default policy: every sarrow propagate the same content to
-  each of its values"""
-function propagate!{T}(carr:: CompArrow,
-       sprtvals::Dict{SubPort, T})::Dict{SubPort, T}
-  propagate!(carr, sprtvals, same_content_propagator)
+"Mapping from `port_id` of `tarr` to abstract values within `abtvals`"
+function tarr_idabvals(tarr::TraceSubArrow, abtvals::AbTraceValues)::IdAbValues
+  # Get IdAbValues from tarr
+  tvals = trace_values(tarr)
+  validids = [prt.port_id for prt in ⬧(deref(tarr)) if tvals[prt.port_id] in keys(abtvals)]
+  IdAbValues(port_id => abtvals[tvals[port_id]] for port_id in validids)
 end
+
+"""
+Propagation
+# Arguments
+- `carr` - the composite arrow to propagate through
+- `val_abval` - any initial values that propagation should be initialized with
+- `tparent` - root of trace (this typicaally set automatically)
+# Returns
+- `val_abval` - mutates and returns val_abval with propagated values
+"""
+function traceprop!(carr::CompArrow,
+                    val_abval::AbTraceValues=AbTraceValues(),
+                    tparent::TraceParent=TraceParent(carr))
+  Time = Int
+  tarrs = inner_trace_arrows(carr)
+  # last time a tarr was applied
+  lastapply = Dict{TraceSubArrow, Time}(zip(tarrs, fill(-1, length(tarrs))))
+  # last time a value was contracted
+  lastmeet = Dict{TraceValue, Time}()
+  for tarr in tarrs, tval in trace_values(tarr)
+    lastmeet[tval] = 0
+  end
+  t::Time = 0
+  # a tarr is ready to to be applied if its values have been `meet`ed more
+  # more recently than the last time it was applied
+  ready(tarr)::Bool = any(value->lastmeet[value] > lastapply[tarr],
+                          trace_values(tarr))
+  while true
+    tarrid = findfirst(ready, tarrs)
+    # Converged
+    tarrid == 0 && break
+    tarr = tarrs[tarrid]
+    lastapply[tarr] = t
+    # An abinterpret abstract evaluator is a function which applies f to concrete domains
+    parr = deref(tarr)
+    @assert isa(parr, PrimArrow)
+
+    # Get IdAbValues from tarr
+    tvals = trace_values(tarr)
+    validids = [prt.port_id for prt in ⬧(deref(tarr)) if tvals[prt.port_id] in keys(val_abval)]
+    idabvals = IdAbValues(port_id => val_abval[tvals[port_id]] for port_id in validids)
+
+    # Do the actual abstract interpretation
+    idabvals = cycle_abinterprets(parr, idabvals)
+
+    # Update `val_abval` with abstract values from idabvals
+    # and update times
+    for (port_id, idabval) in idabvals
+      tval = tvals[port_id]
+      if tval ∉ keys(val_abval) || val_abval[tval] != idabval
+        val_abval[tval] = idabval
+        lastmeet[tval] = t + 1
+      end
+    end
+    t = t + 1
+  end
+  val_abval
+end
+
+"Convenience for specifying abstraact values for subports on root"
+function traceprop!(carr::CompArrow,
+                    sprtprp::Dict{SubPort, AbValues})
+  tparent = TraceParent(carr)
+  val_abval = Dict{TraceValue, AbValues}(TraceValue(tparent, sprt) => props for (sprt, props) in sprtprp)
+  traceprop!(carr, val_abval)
+end
+
+@pre traceprop! !isrecursive(carr)
